Add files via upload
Occulation- An occultation is an event that occurs when one object is hidden from the observer by another object that passes between them. 
(Definition from Wikipedia)
The astronomical term “occultation” comes from the Latin word “occultus”, which means “hidden”.
 
At the same time, The two conditions that must be satisfied for an occultation to occur: 
1)The body, closest to the observer must have a larger visible size than the far one, and should cover it completely. 

2) It should have a lower surface brightness.

Today on March 24, the waxing crescent ( a phase of lunar cycle) Moon and Venus shared the same right ascension in the constellation Aries.
After few minutes, the Moon and Venus will make the closest approach. The illuminated lunar disk will pass 0°06' south of the planet. .
Observers will see the lunar disk pass in front of Venus, creating a lunar occultation.