Add files via upload
The first picture is little shady. It's the first time when we noticed it. So, lowering the brightness will help us to see a lot of surface features. 
And the 2nd picture is that closer look.  And a little more magnifying we've our final capture of  Jupiter. 

The 4th picture is of Jupiter with its 4 moons-  Io, Europa, Ganymede and Callisto. It was discovered by Galileo and are also known as Galilean Satellites.